#27 Add webpack.ide-config.js + rename CellsGrid Service by Generator to be less generic
