+++ webpack.ide-config.js
@@ -1,0 +1,9 @@
+const path = require('path')
+
+module.exports = {
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname)
+    }
+  }
+}
